# swagger.json
{
  "swagger" : "2.0",
  "info"    : {
    "version"     : "1.0.0",
    "title"       : "Sovats API",
    "description" : ""
  },
  "schemes": [
		"http"
	],
	"host": "10.10.1.120:666",
...
  "paths": {
...
    "/api/Driver/All": {
      "post": {
        "tags"       : ["Driver"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          {
            "in"     : "body",
            "name"   : "Body",
            "schema" : {
              "type"       : "object",
              "required"   : true,
              "properties" : {
                "firstName"       : { "type": "string" },
                "lastName"        : { "type": "string" },
                "email"           : { "type": "string" },
                "phone"           : { "type": "string" },
                "blocked"         : { "type": "boolean" },
                "balanceFrom"     : { "type": "number" },
                "balanceTo"       : { "type": "number" }
              }
            }
          }
        ],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
...
    "/api/Promotion/Details/{id}": {
      "get": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "in":"path", "name":"id", "type":"string", "required": true }
        ],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Promotion/Create": {
      "post": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "name"        : { "$ref": "#/parameters/Translates" },
              "description" : { "$ref": "#/parameters/Translates" }
            }
          }
        }],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Promotion/ForAdmin": {
      "get": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [{ "$ref": "#/parameters/Authorization" }],
        "responses"  : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Promotion/Edit/{categoryId}": {
      "put": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in":"path", "name":"categoryId", "type":"string", "required": true },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "name"        : { "$ref": "#/parameters/Translates" },
              "description" : { "$ref": "#/parameters/Translates" }
            }
          }
        }]
      }
    },
    "/api/Promotion/Status/{categoryId}": {
      "put": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in":"path", "name":"categoryId", "type":"string", "required": true },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "status" : {"type": "number"}
            }
          }
        }]
      }
    },
    "/api/Promotion/Position": {
      "put": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "from" : {"type": "number"},
              "to"   : {"type": "number"}
            }
          }
        }]
      }
    },
    "/api/Promotion/Upload": {
      "put": {
        "tags"       : ["Promotion"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in":"query", "name":"id", "type":"string", "required": true },
          { "in":"query", "name":"imageType", "type":"number", "required": true }]
      }
    },
...
    "/api/Admin/Details": {
      "get": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [{ "$ref": "#/parameters/Authorization" }],
        "responses"  : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Admin/List": {
      "get": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in": "query", "name": "role", "type": "number", "required": true }
        ],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Admin/Upload/{id}": {
      "put": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in": "path", "name":"id", "type":"string", "required": true },
          { "in": "formData", "name":"imagePath", "type":"file", "required": true }
        ]
      }
    },
    "/api/Admin/ChangePassword": {
      "put": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          {
            "in"     : "body",
            "name"   : "Body",
            "schema" : {
              "type"       : "object",
              "required"   : true,
              "properties" : {
                "oldPassword": {
                  "type"     : "string",
                  "required" : true
                },
                "newPassword": {
                  "type"     : "string",
                  "required" : true
                },
                "newConfirmPassword": {
                  "type"     : "string",
                  "required" : true
                }
              }
            }
          }
        ],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Admin/Edit": {
      "put": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "firstName" : { "type": "string", "required": true },
              "lastName"  : { "type": "string", "required": true },
              "email"     : { "type": "string", "required": true },
              "phone"     : { "type": "string", "required": true }
            }
          }
        }],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Admin/Edit/{id}": {
      "put": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in": "path", "name": "id", "type": "string", "required": true },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "firstName" : { "type": "string", "required": true },
              "lastName"  : { "type": "string", "required": true },
              "email"     : { "type": "string", "required": true },
              "phone"     : { "type": "string", "required": true },
              "role"      : { "type": "number", "required": true }
            }
          }
        }],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Admin/Create": {
      "post": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          {
          "in"     : "body",
          "name"   : "Body",
          "schema" : {
            "type"       : "object",
            "required"   : true,
            "properties" : {
              "firstName" : { "type": "string", "required": true },
              "lastName"  : { "type": "string", "required": true },
              "email"     : { "type": "string", "required": true },
              "phone"     : { "type": "string", "required": true },
              "role"      : { "type": "number", "required": true },
              "password"  : { "type": "string", "required": true }
            }
          }
        }],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
    "/api/Admin/Delete/{id}": {
      "delete": {
        "tags"       : ["Admin"],
        "consumes"   : "application/json",
        "produces"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" },
          { "in": "path", "name": "id", "type": "string", "required": true }
        ],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
...
    "/api/Settings/All": {
      "get": {
        "tags"       : ["Settings"],
        "consumes"   : "application/json",
        "parameters" : [
          { "$ref": "#/parameters/Authorization" }
        ],
        "responses" : { "200": { "description" : "Successfully response object" }}
      }
    },
...
  },
  "parameters": {
    "Authorization": {
      "name" : "Authorization",
      "in"   : "header",
      "type" : "string"
    },
    "Translates": {
      "type"       : "object",
      "properties" : {
        "hy": {
          "type" : "string"
        },
        "ru": {
          "type" : "string"
        },
        "en": {
          "type" : "string"
        }
      }
    },
...
  }
}